pysparksnippets/requirements.txt
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
black==19.10b0
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
databricks-connect==7.1.0
decorator==4.4.2
defusedxml==0.6.0
dependencies==4.0.1
dependency-injector==4.0.0
entrypoints==0.3
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.0.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mypy==0.782
mypy-extensions==0.4.3
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==6.1.4
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
py==1.9.0
py4j==0.10.9
pycparser==2.20
Pygments==2.7.1
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
pyspark-stubs==3.0.0.post1
pytest==6.1.1
python-dateutil==2.8.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
regex==2020.9.27
requests==2.24.0
Send2Trash==1.5.0
six==1.15.0
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.3.0
